requirements.txt
absl-py==1.3.0
aiofiles==22.1.0
aiosqlite==0.18.0
altgraph==0.17
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
astroid==2.3.3
asttokens==2.2.1
astunparse==1.6.3
attrs==20.2.0
Automat==20.2.0
Babel==2.12.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
bleach==6.0.0
blis==0.4.1
boto3==1.14.56
botocore==1.17.56
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.4.5.1
cffi==1.14.2
chardet==3.0.4
charset-normalizer==3.1.0
ChatterBot==1.0.0
chatterbot-corpus==1.2.0
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.3
coloredlogs==14.0
comm==0.1.2
constantly==15.1.0
Corpus==0.4.2
cycler==0.10.0
cymem==2.0.3
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.2
djangorestframework==3.14.0
docutils==0.15.2
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-2.3.1/en_core_web_md-2.3.1.tar.gz
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.8.0
Flask==1.1.2
flatbuffers==22.9.24
fqdn==1.5.1
future==0.18.2
gast==0.3.3
gevent==20.6.2
google-api-core==1.21.0
google-api-python-client==1.10.0
google-auth==1.19.2
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
graphviz==0.20.1
greenlet==0.4.16
grpcio==1.31.0
h5py==2.10.0
httplib2==0.18.1
humanfriendly==8.2
hyperlink==20.0.1
idna==2.9
importlib-metadata==6.1.0
importlib-resources==5.12.0
incremental==17.5.0
inflection==0.5.1
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
isoduration==20.11.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==0.15.1
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.3
jupyter_client==8.0.3
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.20.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
libclang==14.0.6
Markdown==3.2.2
MarkupSafe==2.1.2
mathparse==0.1.2
matplotlib==3.2.2
matplotlib-inline==0.1.6
mccabe==0.6.1
mistune==2.0.5
more-itertools==8.5.0
murmurhash==1.0.2
mysql-connector-python==8.0.30
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.5.6
nltk==3.5
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.23.4
oauthlib==3.1.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==23.0
pandas==1.0.5
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.1.2
Pint==0.15
pipenv==2022.9.8
pkgutil_resolve_name==1.3.10
plac==1.1.3
platformdirs==2.5.2
preshed==3.0.2
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.12.2
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot-ng==2.0.0
pygame==1.9.6
Pygments==2.14.0
PyHamcrest==2.0.2
PyInstaller==3.6
PyJWT==1.7.1
pylint==2.4.4
pymongo==3.11.0
pynput==1.6.8
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.16.0
python-dateutil==2.8.2
python-dotenv==0.14.0
python-json-logger==2.0.7
pytz==2020.1
pywin32==305
pywin32-ctypes==0.2.0
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.1
Quandl==3.5.2
regex==2020.7.14
requests==2.28.2
requests-oauthlib==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.6
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.12.2
selenium==3.141.0
Send2Trash==1.8.0
simplejson==3.17.2
six==1.14.0
sklearn==0.0
sniffio==1.3.0
soupsieve==2.4
spacy==2.3.2
SQLAlchemy==1.2.19
sqlparse==0.4.3
srsly==1.0.2
stack-data==0.6.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==1.1.0
terminado==0.17.1
Theano==1.0.5
thinc==7.4.1
threadpoolctl==2.1.0
tinycss2==1.2.1
tomli==2.0.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.48.2
traitlets==5.9.0
twilio==6.44.0
typing==3.7.4.3
typing_extensions==4.4.0
tzdata==2022.5
uri-template==1.2.0
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.16.5
virtualenv-clone==0.5.7
wasabi==0.8.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==1.0.1
wrapt==1.11.2
xlrd==1.2.0
y-py==0.5.9
ypy-websocket==0.8.2
zipp==3.15.0
zope.event==4.4
zope.interface==5.1.0
